requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
attrs==25.3.0
backports.tarfile==1.2.0
bcrypt==4.3.0
boto3==1.40.11
botocore==1.40.11
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-log==0.4.0
colorama==0.4.6
cryptography==45.0.6
deprecation==2.1.0
diff-match-patch==20230430
distro==1.9.0
dnspython==2.7.0
docutils==0.22
dotty-dict==1.3.1
ecdsa==0.19.1
email_validator==2.2.0
Faker==25.9.2
fastapi==0.116.1
filelock==3.19.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
gotrue==1.3.1
griffe==0.36.9
grpcio==1.74.0
guardrails-ai==0.6.6
guardrails-api-client==0.4.0a1
guardrails_hub_types==0.0.4
h11==0.14.0
h2==4.2.0
hf-xet==1.1.7
hpack==4.1.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
invoke==1.7.3
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
langchain-core==0.3.74
langsmith==0.4.14
lark==1.2.2
litellm==1.75.6
lxml==4.9.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
milvus-lite==2.5.1
mmh3==5.2.0
more-itertools==10.7.0
multidict==6.6.4
nh3==0.3.0
numpy==2.3.2
openai==1.99.9
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pkginfo==1.12.1.2
postgrest==0.10.7
propcache==0.3.2
protobuf==6.32.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydash==7.0.7
Pygments==2.19.2
PyJWT==2.10.1
pymilvus==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-gitlab==3.15.0
python-jose==3.5.0
python-multipart==0.0.20
python-semantic-release==7.33.2
pytz==2025.2
PyYAML==6.0.2
readme_renderer==44.0
realtime==1.0.6
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.27.0
rsa==4.9.1
rstr==3.2.2
s3transfer==0.13.1
semver==2.13.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supabase_auth==2.12.3
supabase_functions==0.10.1
supafunc==0.2.2
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
tomlkit==0.13.3
tqdm==4.67.1
twine==3.8.0
typer==0.15.4
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.35.0
webcolors==24.11.1
websockets==12.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
